pyproject.toml
[project]
name = "silicon-schema"
version = "0.1.0"
description = "Build tools for SiliconSchema chip data repository"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.scripts]
build-schema = "silicon_schema.build:main"
validate-schema = "silicon_schema.validate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/silicon_schema"]
